+++ R/ggchord.R
@@ -56,7 +56,7 @@ globalVariables(c(
 #'   gene_data = gene_data_example
 #' ) +
 #'   geom_seq() +
-#'   geom_ribbon() +
+#'   geom_link_ribbon() +
 #'   geom_gene()
 #' print(p)
 #'