+++ R/layout-export.R
@@ -24,7 +24,7 @@
 #' data(seq_data_example)
 #' data(ribbon_data_example)
 #' p <- ggchord(seq_data_example, ribbon_data_example) +
-#'   geom_seq() + geom_ribbon()
+#'   geom_seq() + geom_link_ribbon()
 #' exported <- export_ggchord_layout(p, include = c("seq", "ribbon"))
 #' names(exported)
 export_ggchord_layout <- function(